src/CryptoDataManager.ts
import { TokenInfo, Tokenomics, SocialSentiment, TokenReport } from './interfaces';

export class CryptoDataManager {
    async fetchTokenInfo(tokenName: string): Promise<TokenInfo> {
        // Mock API call to Dexscreener and Birdseye
        const tokenInfo: TokenInfo = {
            name: tokenName,
            price: 123.45,
            volume: 67890,
        };
        return tokenInfo;
    }

    async fetchTokenomics(tokenName: string): Promise<Tokenomics> {
        // Mock API call to Cryptorank
        const tokenomics: Tokenomics = {
            marketCap: 1000000,
            totalSupply: 500000,
            circulatingSupply: 300000,
        };
        return tokenomics;
    }

    async fetchSocialSentiment(tokenName: string): Promise<SocialSentiment> {
        // Mock API call to Twitter API, TOTO, or Apify
        const socialSentiment: SocialSentiment = {
            tweets: [
                { body: 'Token', username: 'Himanshu', date: new Date('2024-07-19').toString() },
                { body: 'Project!', username: 'Rohit', date: new Date('2024-07-19').toString() },
            ],
        };
        return socialSentiment;
    }

    async generateReport(tokenName: string): Promise<TokenReport> {
        const [tokenInfo, tokenomics, socialSentiment] = await Promise.all([
            this.fetchTokenInfo(tokenName),
            this.fetchTokenomics(tokenName),
            this.fetchSocialSentiment(tokenName),
        ]);

        const report: TokenReport = {
            tokenInfo,
            tokenomics,
            socialSentiment,
        };

        return report;
    }
}
